mvvm part 2 added
